+++ DatadogCore/Sources/Core/Upload/BackgroundTaskCoordinator.swift
@@ -7,10 +7,11 @@
 import Foundation
 
 internal protocol BackgroundTaskCoordinator {
-    func registerBackgroundTask() -> UUID
-    @discardableResult
-    func endBackgroundTaskIfActive(_ uuid: UUID) -> Bool
+    func registerBackgroundTask() -> BackgroundTaskIdentifier?
+    func endBackgroundTaskIfActive(_ backgroundTaskIdentifier: BackgroundTaskIdentifier)
 }
+
+internal typealias BackgroundTaskIdentifier = Int
 
 #if canImport(UIKit)
 import UIKit
@@ -18,24 +19,12 @@ import UIKit
 /// The `BackgroundTaskCoordinator` class provides an abstraction for managing background tasks and includes methods for registering and ending background tasks.
 /// It also serves as a useful abstraction for testing purposes.
 internal class UIKitBackgroundTaskCoordinator: BackgroundTaskCoordinator {
-    internal var tasks: [UUID: UIBackgroundTaskIdentifier] = [:]
-
-    internal func registerBackgroundTask() -> UUID {
-        let uuid = UUID()
-        tasks[uuid] = UIApplication.dd.managedShared?.beginBackgroundTask { [weak self] in
-            self?.endBackgroundTaskIfActive(uuid)
-        }
-        return uuid
+    internal func registerBackgroundTask() -> BackgroundTaskIdentifier? {
+        return UIApplication.dd.managedShared?.beginBackgroundTask().rawValue
     }
 
-    @discardableResult
-    internal func endBackgroundTaskIfActive(_ uuid: UUID) -> Bool {
-        if let backgroundTask = tasks[uuid], backgroundTask != .invalid {
-            UIApplication.dd.managedShared?.endBackgroundTask(backgroundTask)
-            tasks.removeValue(forKey: uuid)
-            return true
-        }
-        return false
+    func endBackgroundTaskIfActive(_ backgroundTaskIdentifier: BackgroundTaskIdentifier) {
+        UIApplication.dd.managedShared?.endBackgroundTask(.init(rawValue: backgroundTaskIdentifier))
     }
 }
 #endif