+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [IMPROVEMENT] Upgrade to PLCrashReporter 1.11.1.
 - [FEATURE] Report session sample rate to the backend with RUM events. See [#1410][]
 - [IMPROVEMENT] Expose Session Replay to Objective-C. see [#1419][]
+- [IMPROVEMENT] Add UIBackgroundTask for uploading jobs. See [#1412][]
 
 # 2.0.0 / 31-07-2023
 
@@ -497,6 +498,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1410]: https://github.com/DataDog/dd-sdk-ios/pull/1410
 [#1413]: https://github.com/DataDog/dd-sdk-ios/pull/1413
 [#1419]: https://github.com/DataDog/dd-sdk-ios/pull/1419
+[#1412]: https://github.com/DataDog/dd-sdk-ios/pull/1412
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu